@@ -6,41 +6,41 @@
     "build": "dotenv -- cross-var vite build --mode production --base=%VITE_PATH_PREFIX%",
     "serve": "vite preview"
   },
+  "type": "module",
   "dependencies": {
   },
   "devDependencies": {
-    "@heroicons/vue": "^2.0.18",
-    "vue": "^3.3.10",
+    "@heroicons/vue": "^2.1.1",
+    "vue": "^3.4.8",
     "vue-router": "^4.2.5",
     "vue-slicksort": "^2.0.5",
-    "@tiptap/pm": "^2.1.13",
-    "@tiptap/vue-3": "^2.1.13",
-    "@tiptap/extension-document": "^2.1.13",
-    "@tiptap/extension-text": "^2.1.13",
-    "@tiptap/extension-paragraph": "^2.1.13",
-    "@tiptap/extension-hard-break": "^2.1.13",
-    "@tiptap/extension-bold": "^2.1.13",
-    "@tiptap/extension-italic": "^2.1.13",
-    "@tiptap/extension-bullet-list": "^2.1.13",
-    "@tiptap/extension-ordered-list": "^2.1.13",
-    "@tiptap/extension-list-item": "^2.1.13",
-    "@tiptap/extension-history": "^2.1.13",
-    "@tiptap/extension-link": "^2.1.13",
-    "@tiptap/extension-image": "^2.1.13",
-    "@vueuse/core": "^10.7.0",
+    "@tiptap/pm": "^2.1.15",
+    "@tiptap/vue-3": "^2.1.15",
+    "@tiptap/extension-document": "^2.1.15",
+    "@tiptap/extension-text": "^2.1.15",
+    "@tiptap/extension-paragraph": "^2.1.15",
+    "@tiptap/extension-hard-break": "^2.1.15",
+    "@tiptap/extension-bold": "^2.1.15",
+    "@tiptap/extension-italic": "^2.1.15",
+    "@tiptap/extension-bullet-list": "^2.1.15",
+    "@tiptap/extension-ordered-list": "^2.1.15",
+    "@tiptap/extension-list-item": "^2.1.15",
+    "@tiptap/extension-history": "^2.1.15",
+    "@tiptap/extension-link": "^2.1.15",
+    "@tiptap/extension-image": "^2.1.15",
+    "@vueuse/core": "^10.7.1",
     "@tailwindcss/forms": "^0.5.7",
-    "@vitejs/plugin-vue": "^4.5.1",
+    "@vitejs/plugin-vue": "^5.0.3",
     "@vue/cli-service": "^5.0.8",
-    "@vue/compiler-sfc": "^3.3.10",
+    "@vue/compiler-sfc": "^3.4.8",
     "autoprefixer": "^10.4.16",
-    "postcss": "^8.4.32",
-    "tailwindcss": "^3.3.6",
+    "postcss": "^8.4.33",
+    "tailwindcss": "^3.4.1",
     "@tailwindcss/typography": "^0.5.10",
-    "rustywind": "^0.20.0",
-    "vite": "^5.0.5",
-    "vx-vue": "^0.7.5",
+    "rustywind": "^0.21.0",
+    "vite": "^5.0.11",
+    "vx-vue": "^0.7.7",
     "dotenv-cli": "^7.3.0",
-    "cross-var": "^1.1.0",
-    "@niku/vite-env-caster": "^1.0.1"
+    "cross-var": "^1.1.0"
   }
 }
