@@ -70,23 +70,23 @@
       <div class="flex justify-end items-center space-x-1">
         <template v-if="slotProps.row.isFolder">
           <button class="icon-link tooltip" data-tooltip="Bearbeiten" @click="fm.editFolder(slotProps.row)">
-            <pencil-square-icon class="w-5 h-5" />
+            <pencil-square-icon class="size-5" />
           </button>
           <button class="icon-link tooltip" data-tooltip="Löschen" @click="fm.delFolder(slotProps.row)">
-            <trash-icon class="w-5 h-5" />
+            <trash-icon class="size-5" />
           </button>
         </template>
         <template v-else>
           <button class="icon-link tooltip" data-tooltip="Bearbeiten" @click="fm.editFile(slotProps.row)">
-            <pencil-square-icon class="w-5 h-5" />
+            <pencil-square-icon class="size-5" />
           </button>
           <button class="flex items-center icon-link tooltip" data-tooltip="Verschieben" @click="fm.moveFile(slotProps.row)">
-            <document-minus-icon class="w-5 h-5"/>
+            <document-minus-icon class="size-5"/>
             <play-icon class="w-3 h-3" />
-            <document-plus-icon class="w-5 h-5"/>
+            <document-plus-icon class="size-5"/>
           </button>
           <button class="icon-link tooltip" data-tooltip="Löschen" @click="fm.delFile(slotProps.row)">
-            <trash-icon class="w-5 h-5" />
+            <trash-icon class="size-5" />
           </button>
         </template>
       </div>
